requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
networkx>=3.2.1
plotly>=5.20.0
google-genai